SOCINT-311 Refactor Utils Split out externals calls and validators from utils.py
