+++ app/utils.py
@@ -1,33 +1,16 @@
-import string
-import re
-import jwt
-
 from aiohttp.client_exceptions import (ClientResponseError)
-from .exceptions import InactiveCaseError, InvalidDataError, InvalidDataErrorWelsh, \
-    TooManyRequestsEQLaunch
 from aiohttp.web import HTTPFound
-from datetime import datetime
-from pytz import timezone, utc
-from unicodedata import normalize
-
+from pytz import timezone
 from sdc.crypto.encrypter import encrypt
+
+from .exceptions import TooManyRequestsEQLaunch
 from .eq import EqPayloadConstructor
-from .flash import flash
-from .request import RetryRequest
+from .service_calls.rhsvc import RHSvcEQLaunch
+
 from structlog import get_logger
 
 logger = get_logger('respondent-home')
 
-OBSCURE_WHITESPACE = (
-    '\u180E'  # Mongolian vowel separator
-    '\u200B'  # zero width space
-    '\u200C'  # zero width non-joiner
-    '\u200D'  # zero width joiner
-    '\u2060'  # word joiner
-    '\uFEFF'  # zero width non-breaking space
-)
-
-uk_prefix = '44'
 uk_zone = timezone('Europe/London')
 
 
@@ -37,31 +20,6 @@ class View:
     valid_user_journeys = r'{user_journey:\bstart|request\b}'
     page_title_error_prefix_en = 'Error: '
     page_title_error_prefix_cy = 'Gwall: '
-
-    @staticmethod
-    def single_client_ip(request):
-        if request['client_ip']:
-            client_ip = request['client_ip']
-            single_ip_validation_pattern = re.compile(r'^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$')
-            if client_ip.count(',') > 1:
-                single_ip_value = client_ip.split(', ', -1)[-3]
-                if single_ip_validation_pattern.fullmatch(single_ip_value):
-                    single_ip = single_ip_value
-                else:
-                    logger.warn('clientIP failed validation. Provided IP - ' + client_ip,
-                                client_id=request['client_id'],
-                                trace=request['trace'])
-                    single_ip = ''
-            else:
-                logger.warn('clientIP failed validation. Provided IP - ' + client_ip,
-                            client_id=request['client_id'],
-                            trace=request['trace'])
-                single_ip = ''
-        elif request.headers.get('Origin', None) and 'localhost' in request.headers.get('Origin', None):
-            single_ip = '127.0.0.1'
-        else:
-            single_ip = ''
-        return single_ip
 
     @staticmethod
     def log_entry(request, endpoint):
@@ -115,41 +73,18 @@ class View:
 
         return link
 
-    @staticmethod
-    async def _make_request(request,
-                            method,
-                            url,
-                            auth=None,
-                            headers=None,
-                            request_json=None,
-                            return_json=False):
-        """
-        :param request: The AIOHTTP user request, used for logging and app access
-        :param method: The HTTP verb
-        :param url: The target URL
-        :param auth: Authorization
-        :param headers: Any needed headers as a python dictionary
-        :param request_json: JSON payload to pass as request data
-        :param return_json: If True, the response JSON will be returned
-        """
-        retry_request = RetryRequest(request, method, url, auth, headers, request_json, return_json)
-        return await retry_request.make_request()
 
+class LaunchEQ:
     @staticmethod
-    def validate_case(case_json):
-        if not case_json.get('active', False):
-            raise InactiveCaseError()
-
-    @staticmethod
-    async def call_questionnaire(request, case, attributes, app, adlocation):
-        eq_payload = await EqPayloadConstructor(case, attributes, app).build()
+    async def call_questionnaire(request, uac_context, attributes, app):
+        eq_payload = await EqPayloadConstructor(uac_context, attributes, app).build()
 
         token = encrypt(eq_payload,
                         key_store=app['key_store'],
                         key_purpose='authentication')
 
         try:
-            await RHService.post_surveylaunched(request, case, adlocation)
+            await RHSvcEQLaunch.post_survey_launched(request, uac_context)
         except ClientResponseError as ex:
             if ex.status == 429:
                 raise TooManyRequestsEQLaunch()
@@ -162,447 +97,8 @@ class View:
         raise HTTPFound(f'{eq_url}/session?token={token}')
 
 
-class ProcessPostcode:
-    postcode_validation_pattern = re.compile(
-        r'^((AB|AL|B|BA|BB|BD|BH|BL|BN|BR|BS|BT|BX|CA|CB|CF|CH|CM|CO|CR|CT|CV|CW|DA|DD|DE|DG|DH|DL|DN|DT|DY|E|EC|EH|EN|EX|FK|FY|G|GL|GY|GU|HA|HD|HG|HP|HR|HS|HU|HX|IG|IM|IP|IV|JE|KA|KT|KW|KY|L|LA|LD|LE|LL|LN|LS|LU|M|ME|MK|ML|N|NE|NG|NN|NP|NR|NW|OL|OX|PA|PE|PH|PL|PO|PR|RG|RH|RM|S|SA|SE|SG|SK|SL|SM|SN|SO|SP|SR|SS|ST|SW|SY|TA|TD|TF|TN|TQ|TR|TS|TW|UB|W|WA|WC|WD|WF|WN|WR|WS|WV|YO|ZE)(\d[\dA-Z]?[ ]?\d[ABD-HJLN-UW-Z]{2}))$'  # NOQA
-    )
-
-    @staticmethod
-    def validate_postcode(postcode, locale):
-
-        for character in string.whitespace + OBSCURE_WHITESPACE:
-            postcode = postcode.replace(character, '')
-
-        postcode = postcode.upper()
-        postcode = normalize('NFKD', postcode).encode('ascii', 'ignore').decode('utf8')
-
-        if len(postcode) == 0:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch god post", 'empty')
-            else:
-                raise InvalidDataError('Enter a postcode', 'empty')
-
-        if not postcode.isalnum():
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch god post dilys yn y Deyrnas Unedig")
-            else:
-                raise InvalidDataError('Enter a valid UK postcode')
-
-        if len(postcode) < 5:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch god post dilys yn y Deyrnas Unedig")
-            else:
-                raise InvalidDataError('Enter a valid UK postcode')
-
-        if len(postcode) > 7:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch god post dilys yn y Deyrnas Unedig")
-            else:
-                raise InvalidDataError('Enter a valid UK postcode')
-
-        if not ProcessPostcode.postcode_validation_pattern.fullmatch(postcode):
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch god post dilys yn y Deyrnas Unedig")
-            else:
-                raise InvalidDataError('Enter a valid UK postcode')
-
-        postcode_formatted = postcode[:-3] + ' ' + postcode[-3:]
-
-        return postcode_formatted
-
-
-class ProcessMobileNumber:
-
-    @staticmethod
-    def normalise_phone_number(number, locale):
-
-        for character in string.whitespace + OBSCURE_WHITESPACE + '()-+':
-            number = number.replace(character, '')
-
-        try:
-            list(map(int, number))
-        except ValueError:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch rif ffôn symudol yn y Deyrnas Unedig mewn fformat dilys, "
-                                            "er enghraifft, 07700 900345 neu +44 7700 900345", message_type='invalid')
-            else:
-                raise InvalidDataError('Enter a UK mobile number in a valid format, for example, '
-                                       '07700 900345 or +44 7700 900345', message_type='invalid')
-
-        return number.lstrip('0')
-
-    @staticmethod
-    def validate_uk_mobile_phone_number(number, locale):
-
-        number = ProcessMobileNumber.normalise_phone_number(number, locale).lstrip(uk_prefix).lstrip('0')
-
-        if len(number) == 0:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch eich rhif ffôn symudol", message_type='empty')
-            else:
-                raise InvalidDataError('Enter your mobile number', message_type='empty')
-
-        if not number.startswith('7'):
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch rif ffôn symudol yn y Deyrnas Unedig mewn fformat dilys, "
-                                            "er enghraifft, 07700 900345 neu +44 7700 900345", message_type='invalid')
-            else:
-                raise InvalidDataError('Enter a UK mobile number in a valid format, for example, '
-                                       '07700 900345 or +44 7700 900345', message_type='invalid')
-
-        if len(number) > 10:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch rif ffôn symudol yn y Deyrnas Unedig mewn fformat dilys, "
-                                            "er enghraifft, 07700 900345 neu +44 7700 900345", message_type='invalid')
-            else:
-                raise InvalidDataError('Enter a UK mobile number in a valid format, for example, '
-                                       '07700 900345 or +44 7700 900345', message_type='invalid')
-
-        if len(number) < 10:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Rhowch rif ffôn symudol yn y Deyrnas Unedig mewn fformat dilys, "
-                                            "er enghraifft, 07700 900345 neu +44 7700 900345", message_type='invalid')
-            else:
-                raise InvalidDataError('Enter a UK mobile number in a valid format, for example, '
-                                       '07700 900345 or +44 7700 900345', message_type='invalid')
-
-        return '{}{}'.format(uk_prefix, number)
-
-
-class ProcessName:
-
-    @staticmethod
-    def validate_name(request, data, display_region, child=False):
-
-        name_valid = True
-        form_first_name = data.get('name_first_name')
-        form_last_name = data.get('name_last_name')
-
-        if (not form_first_name) or (len(form_first_name.strip()) == 0):
-            if child:
-                flash(request, {'text': "Enter your child's first name", 'level': 'ERROR', 'type': 'NAME_ENTER_ERROR',
-                                'field': 'error_first_name', 'value_first_name': form_first_name,
-                                'value_last_name': form_last_name})
-            else:
-                if display_region == 'cy':
-                    flash(request, {'text': "Rhowch eich enw cyntaf", 'level': 'ERROR', 'type': 'NAME_ENTER_ERROR',
-                                    'field': 'error_first_name', 'value_first_name': form_first_name,
-                                    'value_last_name': form_last_name})
-                else:
-                    flash(request, {'text': "Enter your first name", 'level': 'ERROR', 'type': 'NAME_ENTER_ERROR',
-                                    'field': 'error_first_name', 'value_first_name': form_first_name,
-                                    'value_last_name': form_last_name})
-            name_valid = False
-
-        elif len(form_first_name) > 35:
-            if display_region == 'cy':
-                flash(request, {'text': "Rydych wedi defnyddio gormod o nodau. Rhowch hyd at 35 o nodau",
-                                'level': 'ERROR', 'type': 'NAME_ENTER_ERROR', 'field': 'error_first_name_len',
-                                'value_first_name': form_first_name, 'value_last_name': form_last_name})
-            else:
-                flash(request, {'text': 'You have entered too many characters. Enter up to 35 characters',
-                                'level': 'ERROR', 'type': 'NAME_ENTER_ERROR', 'field': 'error_first_name_len',
-                                'value_first_name': form_first_name, 'value_last_name': form_last_name})
-            name_valid = False
-
-        if (not form_last_name) or (len(form_last_name.strip()) == 0):
-            if child:
-                flash(request, {'text': "Enter your child's last name", 'level': 'ERROR', 'type': 'NAME_ENTER_ERROR',
-                                'field': 'error_last_name', 'value_first_name': form_first_name,
-                                'value_last_name': form_last_name})
-            else:
-                if display_region == 'cy':
-                    flash(request, {'text': "Rhowch eich cyfenw", 'level': 'ERROR', 'type': 'NAME_ENTER_ERROR',
-                                    'field': 'error_last_name', 'value_first_name': form_first_name,
-                                    'value_last_name': form_last_name})
-                else:
-                    flash(request, {'text': "Enter your last name", 'level': 'ERROR', 'type': 'NAME_ENTER_ERROR',
-                                    'field': 'error_last_name', 'value_first_name': form_first_name,
-                                    'value_last_name': form_last_name})
-            name_valid = False
-
-        elif len(form_last_name) > 35:
-            if display_region == 'cy':
-                flash(request, {'text': "Rydych wedi defnyddio gormod o nodau. Rhowch hyd at 35 o nodau",
-                                'level': 'ERROR', 'type': 'NAME_ENTER_ERROR', 'field': 'error_last_name_len',
-                                'value_first_name': form_first_name, 'value_last_name': form_last_name})
-            else:
-                flash(request, {'text': 'You have entered too many characters. Enter up to 35 characters',
-                                'level': 'ERROR', 'type': 'NAME_ENTER_ERROR', 'field': 'error_last_name_len',
-                                'value_first_name': form_first_name, 'value_last_name': form_last_name})
-            name_valid = False
-
-        return name_valid
-
-
-class ProcessDOB:
-    @staticmethod
-    def validate_dob(data):
-        form_day = data.get('day')
-        form_month = data.get('month')
-        form_year = data.get('year')
-
-        try:
-            date = datetime(int(form_year), int(form_month), int(form_day)).date()
-            return date
-        except ValueError:
-            raise InvalidDataError('invalid dob', message_type='invalid')
-
-    @staticmethod
-    def format_dob(date_value):
-        unformatted_date = datetime.strptime(date_value, '%Y-%m-%d')
-        formatted_date = unformatted_date.strftime('%d %B %Y')
-        return formatted_date
-
-
-class ProcessEmailAddress:
-    email_validation_pattern = re.compile(
-        r'(^[^@\s]+@[^@\s]+\.[^@\s]+$)'
-    )
-
-    @staticmethod
-    def validate_email(email, locale):
-        if len(email.strip()) == 0:
-            if locale == 'cy':
-                raise InvalidDataErrorWelsh("Enter an email address", message_type='empty')
-            else:
-                raise InvalidDataError('Enter an email address', message_type='empty')
-        else:
-            if ProcessEmailAddress.email_validation_pattern.fullmatch(email):
-                return email
-            else:
-                if locale == 'cy':
-                    raise InvalidDataErrorWelsh(
-                        "Enter an email address in a valid format, for example, name@example.com",
-                        message_type='invalid')
-                else:
-                    raise InvalidDataError(
-                        'Enter an email address in a valid format, for example, name@example.com',
-                        message_type='invalid')
-
-
 class FlashMessage:
-
     @staticmethod
     def generate_flash_message(text, level, message_type, field):
         json_return = {'text': text, 'level': level, 'type': message_type, 'field': field}
         return json_return
-
-
-class AddressIndex(View):
-
-    @staticmethod
-    def generate_jwt(request):
-        key = request.app['ADDRESS_INDEX_SVC_KEY']
-        token = jwt.encode({}, key, algorithm="HS256")
-        return token
-
-    @staticmethod
-    async def get_postcode_return(request, postcode, display_region):
-        postcode_return = await AddressIndex.get_ai_postcode(request, postcode)
-
-        address_options = []
-
-        if display_region == 'cy':
-            cannot_find_text = 'Ni allaf ddod o hyd i fy nghyfeiriad'
-        else:
-            cannot_find_text = 'I cannot find my address'
-
-        for singleAddress in postcode_return['response']['addresses']:
-            address_options.append({
-                'value': singleAddress['uprn'],
-                'label': {
-                    'text': singleAddress['formattedAddress']
-                },
-                'id': singleAddress['uprn']
-            })
-
-        address_options.append({
-            'value': 'xxxx',
-            'label': {
-                'text': cannot_find_text
-            },
-            'id': 'xxxx'
-        })
-
-        address_content = {
-            'postcode': postcode,
-            'addresses': address_options,
-            'total_matches': postcode_return['response']['total']
-        }
-
-        return address_content
-
-    @staticmethod
-    async def get_ai_postcode(request, postcode):
-        ai_svc_url = request.app['ADDRESS_INDEX_SVC_URL']
-        ai_epoch = request.app['ADDRESS_INDEX_EPOCH']
-        token = AddressIndex.generate_jwt(request)
-        url = f'{ai_svc_url}/addresses/rh/postcode/{postcode}?limit=5000&epoch={ai_epoch}'
-        headers = {'Authorization': 'Bearer ' + token}
-        return await View._make_request(request,
-                                        'GET',
-                                        url,
-                                        headers=headers,
-                                        return_json=True)
-
-    @staticmethod
-    async def get_ai_uprn(request, uprn):
-        ai_svc_url = request.app['ADDRESS_INDEX_SVC_URL']
-        ai_epoch = request.app['ADDRESS_INDEX_EPOCH']
-        url = f'{ai_svc_url}/addresses/rh/uprn/{uprn}?addresstype=paf&epoch={ai_epoch}'
-        headers = {'Authorization': 'Bearer ' + AddressIndex.generate_jwt(request)}
-        return await View._make_request(request,
-                                        'GET',
-                                        url,
-                                        headers=headers,
-                                        return_json=True)
-
-
-class RHService(View):
-
-    @staticmethod
-    async def get_case_by_uprn(request, uprn):
-        rhsvc_url = request.app['RHSVC_URL']
-        return await View._make_request(request,
-                                        'GET',
-                                        f'{rhsvc_url}/cases/uprn/{uprn}',
-                                        return_json=True)
-
-    @staticmethod
-    async def get_fulfilment(request, region,
-                             delivery_channel, product_group, individual):
-        rhsvc_url = request.app['RHSVC_URL']
-        url = f'{rhsvc_url}/fulfilments?caseType=HH&region={region}&deliveryChannel={delivery_channel}' \
-              f'&productGroup={product_group}&individual={individual}'
-        return await View._make_request(request,
-                                        'GET',
-                                        url,
-                                        return_json=True)
-
-    @staticmethod
-    async def request_fulfilment_sms(request, case_id, tel_no, fulfilment_code_array):
-        rhsvc_url = request.app['RHSVC_URL']
-        fulfilment_json = {
-            'caseId': case_id,
-            'telNo': tel_no,
-            'fulfilmentCodes': fulfilment_code_array,
-            'dateTime': datetime.now(utc).isoformat(),
-            'clientIP': View.single_client_ip(request)
-        }
-        url = f'{rhsvc_url}/cases/{case_id}/fulfilments/sms'
-        return await View._make_request(request,
-                                        'POST',
-                                        url,
-                                        auth=request.app['RHSVC_AUTH'],
-                                        request_json=fulfilment_json)
-
-    @staticmethod
-    async def request_fulfilment_post(request, case_id, first_name, last_name, fulfilment_code_array, title=None):
-        rhsvc_url = request.app['RHSVC_URL']
-        fulfilment_json = {
-            'caseId': case_id,
-            'title': title,
-            'forename': first_name,
-            'surname': last_name,
-            'fulfilmentCodes': fulfilment_code_array,
-            'dateTime': datetime.now(utc).isoformat(),
-            'clientIP': View.single_client_ip(request)
-        }
-        url = f'{rhsvc_url}/cases/{case_id}/fulfilments/post'
-        return await View._make_request(request,
-                                        'POST',
-                                        url,
-                                        auth=request.app['RHSVC_AUTH'],
-                                        request_json=fulfilment_json)
-
-    @staticmethod
-    async def request_fulfilment_email(request, case_id, email, fulfilment_code_array):
-        rhsvc_url = request.app['RHSVC_URL']
-        fulfilment_json = {
-            'caseId': case_id,
-            'email': email,
-            'fulfilmentCodes': fulfilment_code_array,
-            'dateTime': datetime.now(utc).isoformat(),
-            'clientIP': View.single_client_ip(request)
-        }
-        url = f'{rhsvc_url}/cases/{case_id}/fulfilments/email'
-        return await View._make_request(request,
-                                        'POST',
-                                        url,
-                                        auth=request.app['RHSVC_AUTH'],
-                                        request_json=fulfilment_json)
-
-    @staticmethod
-    async def get_uac_details(request):
-        uac_hash = request['uac_hash']
-        logger.info('making get request for uac',
-                    client_ip=request['client_ip'],
-                    client_id=request['client_id'],
-                    trace=request['trace'],
-                    uac_hash=uac_hash)
-        rhsvc_url = request.app['RHSVC_URL']
-        return await View._make_request(request,
-                                        'GET',
-                                        f'{rhsvc_url}/uacs/{uac_hash}',
-                                        auth=request.app['RHSVC_AUTH'],
-                                        return_json=True)
-
-    @staticmethod
-    async def post_surveylaunched(request, case, adlocation):
-        if not adlocation:
-            adlocation = ''
-        launch_json = {
-            'questionnaireId': case['qid'],
-            'caseId': case['collectionCase']['caseId'],
-            'agentId': adlocation,
-            'clientIP': View.single_client_ip(request)
-        }
-        rhsvc_url = request.app['RHSVC_URL']
-        return await View._make_request(request,
-                                        'POST',
-                                        f'{rhsvc_url}/surveyLaunched',
-                                        auth=request.app['RHSVC_AUTH'],
-                                        request_json=launch_json)
-
-    @staticmethod
-    async def post_webform(request, form_data):
-        form_json = {
-            'category': form_data['category'],
-            'region': form_data['region'],
-            'language': form_data['language'],
-            'name': form_data['name'],
-            'description': form_data['description'],
-            'email': form_data['email'],
-            'clientIP': View.single_client_ip(request)
-        }
-        rhsvc_url = request.app['RHSVC_URL']
-        return await View._make_request(request,
-                                        'POST',
-                                        f'{rhsvc_url}/webform',
-                                        auth=request.app['RHSVC_AUTH'],
-                                        request_json=form_json)
-
-    @staticmethod
-    async def register_new_case(request, data):
-        new_case_json = {
-            'schoolId': '1',  # Dummy value - no way to source currently
-            'schoolName': data['school_name'],
-            'consentGivenTest': 'true',  # Fixed value
-            'consentGivenSurvey': 'true',  # Fixed value
-            'firstName': data['parent_first_name'],
-            # Unable to submit parent_middle_names
-            'lastName': data['parent_last_name'],
-            'childFirstName': data['child_first_name'],
-            'childMiddleNames': data['child_middle_names'],
-            'childLastName': data['child_last_name'],
-            'childDob': data['child_dob'],
-            'mobileNumber': data['mobile_number'],
-            'emailAddress': 'a.b@c.com'  # Dummy value - not required to be captured currently
-        }
-        rhsvc_url = request.app['RHSVC_URL']
-        return await View._make_request(request,
-                                        'POST',
-                                        f'{rhsvc_url}/cases/new',
-                                        auth=request.app['RHSVC_AUTH'],
-                                        request_json=new_case_json)
