+++ tests/unit/helpers.py
@@ -350,7 +350,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_enter_address(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.service_calls.aims.AimsPostcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
@@ -368,7 +368,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_enter_address_finder(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, mock.patch(
-                'app.utils.RHService.get_case_by_uprn') as mocked_get_case_by_uprn:
+                'app.service_calls.rhsvc.RHSvcCases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             if region == 'W':
                 mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_w
             else:
@@ -389,7 +389,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_enter_address_input_returns_no_results(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.service_calls.aims.AimsPostcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_no_results
 
             response = await self.client.request('POST',
@@ -450,7 +450,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_address_no_selection_made(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.service_calls.aims.AimsPostcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
@@ -467,8 +467,8 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_address(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
-                'app.utils.RHService.get_case_by_uprn') as mocked_get_case_by_uprn:
+                mock.patch('app.service_calls.aims.AimsPostcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                'app.service_calls.rhsvc.RHSvcCases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             if region == 'W':
                 mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_w
@@ -491,7 +491,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_address_no_case(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_uprn') as mocked_get_ai_uprn, \
+                mock.patch('app.service_calls.aims.AimsUprn.get_ai_uprn') as mocked_get_ai_uprn, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_get_case_by_uprn:
 
             mocked_get_case_by_uprn.get(self.rhsvc_cases_by_uprn_url + self.selected_uprn, status=404)
@@ -572,8 +572,8 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_invalid(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
-                'app.utils.RHService.get_case_by_uprn') as mocked_get_case_by_uprn:
+                mock.patch('app.service_calls.aims.AimsPostcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                'app.service_calls.rhsvc.RHSvcCases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_e
 
@@ -588,8 +588,8 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_no_selection(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
-                'app.utils.RHService.get_case_by_uprn') as mocked_get_case_by_uprn:
+                mock.patch('app.service_calls.aims.AimsPostcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                'app.service_calls.rhsvc.RHSvcCases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_e
 
@@ -604,7 +604,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_no(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.service_calls.aims.AimsPostcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
@@ -777,9 +777,9 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_text(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, mock.patch(
-                'app.utils.RHService.get_fulfilment'
+                'app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment'
         ) as mocked_get_fulfilment, mock.patch(
-            'app.utils.RHService.request_fulfilment_sms'
+            'app.service_calls.rhsvc.RHSvcFulfilments.request_fulfilment_sms'
         ) as mocked_request_fulfilment_sms:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_multi_sms
@@ -844,7 +844,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_text_error_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.RHService.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_sms
             mocked_aioresponses.post(self.rhsvc_cases_url +
@@ -861,7 +861,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_text_error_429_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.RHService.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_sms
@@ -912,9 +912,9 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_email(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, mock.patch(
-                'app.utils.RHService.get_fulfilment'
+                'app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment'
         ) as mocked_get_fulfilment, mock.patch(
-            'app.utils.RHService.request_fulfilment_email'
+            'app.service_calls.rhsvc.RHSvcFulfilments.request_fulfilment_email'
         ) as mocked_request_fulfilment_email:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_multi_email
@@ -979,7 +979,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_email_error_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.RHService.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_email
             mocked_aioresponses.post(self.rhsvc_cases_url +
@@ -996,7 +996,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_email_error_429_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.RHService.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_email
@@ -1208,8 +1208,9 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_post_input_yes(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.RHService.get_fulfilment') as mocked_get_fulfilment, \
-                mock.patch('app.utils.RHService.request_fulfilment_post') as mocked_request_fulfilment_post:
+                mock.patch('app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.RHSvcFulfilments.request_fulfilment_post'
+                           ) as mocked_request_fulfilment_post:
 
             if display_region == 'cy':
                 mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_multi_post
@@ -1290,7 +1291,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_post_error_from_request_fulfilment(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.RHService.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_post
             mocked_aioresponses.post(self.rhsvc_cases_url +
@@ -1307,7 +1308,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_send_by_post_error_429_from_request_fulfilment_uac(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.RHService.get_fulfilment') as mocked_get_fulfilment, \
+                mock.patch('app.service_calls.rhsvc.RHSvcFulfilments.get_fulfilment') as mocked_get_fulfilment, \
                 aioresponses(passthrough=[str(self.server._root)]) as mocked_aioresponses:
 
             mocked_get_fulfilment.return_value = self.rhsvc_get_fulfilment_single_post
